requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
filelock==3.13.3
flake8==7.0.0
frozenlist==1.4.1
fsspec==2024.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.16.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sniffio==1.3.1
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
boto3==1.34.94
